新增操作权限、用户权限查询

+++ spring-platform-base/spring-platform-model/src/main/java/com/platform/model/entity/user/SysRolePermission.java
@@ -18,12 +18,12 @@ public class SysRolePermission extends BaseEntity {
     /**
      * 角色ID
      */
-    private Integer roleId;
+    private Long roleId;
 
     /**
      * 权限ID
      */
-    private Integer permissionId;
+    private Long permissionId;
 
 
     public static final String ROLE_ID = "role_id";
